Create README.md
Placental abnormalities are the cause of impaired fetal growth and also poor pregnancy outcome (e.g. Preeclampsia) with long lasting consequence for the mother and offspring. The molecular dialogue between the maternal niche and the developing placenta is critical for the function of this organ. Galectin-1 (gal-1), a highly expressed glycan-binding protein at the maternal-fetal interface, orchestrates the maternal adaptation to pregnancy and placenta development. Downregulation or deficiency of gal-1 during pregnancy is associated with the development of PE, however, the maternal- and placental- derived gal-1 contributions to the disease onset is largely unknown. We demonstrate that lack of gal-1 imposes a risk for the PE development in a niche-specific manner and this is accompanied by a placental dysfunction highly influenced by the absence of maternal- derived gal-1. Notably, differential placental glycosylation through the Sda-capped N-glycans dominate the trophoblast invasive capacity triggered by maternal- derived gal-1. Our findings show that gal-1 derived from the maternal niche is essential for a healthy placenta development and indicate that impairment of the gal-1 signaling pathway within the maternal niche could be a molecular cause for maternal cardiovascular maladaptation during pregnancy.